Attempt to fix bug that does not print result
Summary: playRound does not print the result of each round. When attempting to fix this by adding console.log after function, prompt window shows up twice instead.

Attempt: Declared variables humanChoice and computerChoice separately (not within () of playRound) to find out if this was affecting bug. It did not. It seems the issue may be in the if...else statement.