Fixed bug where win-lose message became random
Problem: Logging results using console.log after function getComputerChoice caused the function to be called twice -- once after the function and once in playRound. As a result, the first result printed in the console became different from the actual value used in the match (basically an invisible second match).

Solution:
- Delete console.log after function getComputerChoice
- Add console.log for both functions getHumanChoice and getComputerChoice inside of function playRound to deter repeating any functions
- change computerChoice = getComputerChoice() to computerChoice = getComputerChoice(3), for consistency and prevent future errors